# docs/swagger.yaml
basePath: /v1
definitions:
...
  auth.Project:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
  auth.UserRole:
    enum:
    - owner
    - admin
    - viewer
    type: string
    x-enum-varnames:
    - RoleOwner
    - RoleAdmin
    - RoleViewer
  auth.addMemberRequest:
    properties:
      email:
        type: string
      role:
        $ref: '#/definitions/auth.UserRole'
    required:
    - email
    - role
    type: object
  auth.createAPIKeyRequest:
    properties:
      name:
        maxLength: 128
        minLength: 1
        type: string
      project_id:
        type: string
    required:
    - name
    - project_id
    type: object
  auth.createProjectRequest:
    properties:
      name:
        maxLength: 128
        minLength: 1
        type: string
      slug:
        maxLength: 128
        minLength: 1
        type: string
    required:
    - name
    - slug
    type: object
  auth.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.updateMemberRoleRequest:
    properties:
      role:
        $ref: '#/definitions/auth.UserRole'
    required:
    - role
    type: object
host: localhost:8082
info:
  contact:
    name: BatAudit
  description: Self-hosted audit logging platform. Collect, store and query audit
    events from any service.
  license:
    name: MIT
  title: BatAudit API
  version: "1.0"
paths:
...
  /auth/api-keys:
    get:
      description: Returns all API keys for a project
      parameters:
      - description: Project ID
        in: query
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - api-keys
    post:
      consumes:
      - application/json
      description: Generates a new API key for a project. The raw key is shown only
        once.
      parameters:
      - description: API key data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.createAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - api-keys
  /auth/api-keys/{id}:
    delete:
      description: Marks an API key as inactive — it will no longer be accepted by
        the Writer
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Revoke API key
      tags:
      - api-keys
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      description: Invalidates the current session (client should discard the token)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /auth/me:
    get:
      description: Returns the authenticated user's info from the JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Current user
      tags:
      - auth
  /auth/projects:
    get:
      description: Returns all projects the current user has access to. Owner sees
        all projects.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Creates a new project
      parameters:
      - description: Project data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.createProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - projects
  /auth/projects/{id}/members:
    get:
      description: Returns all members of a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List project members
      tags:
      - members
    post:
      consumes:
      - application/json
      description: Adds a user (by email) to a project with the specified role
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Member data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.addMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add project member
      tags:
      - members
  /auth/projects/{id}/members/{userId}:
    delete:
      description: Removes a user from a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove project member
      tags:
      - members
    patch:
      consumes:
      - application/json
      description: Changes the role of an existing project member
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: New role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.updateMemberRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update member role
      tags:
      - members
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"